edit dr Schedule

+++ Frontend/src/app/admin/doctor-schedule/doctor-schedule.component.ts
@@ -43,16 +43,21 @@ export class DoctorScheduleComponent implements OnInit , OnDestroy {
   ngOnInit(): void {
     this.days=['Saturday','Monday','Tuesday','Wednesday','Thursday','Friday'];
     this.shift=['PM','AM'];
-    this.hours= ["00:00","1:00","2:00","3:00","4:00","5:00","6:00","7:00","8:00","9:00","10:00","11:00"];
+    this.hours= ["1:00","2:00","3:00","4:00","5:00","6:00","7:00","8:00","9:00","10:00","11:00","12:00"];
   }
+
   AddToSchedule(){
     this.showAddToSchedule = true ;
-
   }
+
   AddToDB(){
     if(this.info.day =="" ||this.info.start =="" ||this.info.sShift =="" ||this.info.end =="" ||this.info.eShift =="" ||this.info.maxNo ==""){
         this.errorMessageDuringAdd = true ;
         this.errMsg = 'All fields are required !!';
+    }
+    else if(!this.info.maxNo.match(this.IsNumberPattern)){
+        this.errorMessageDuringAdd = true ;
+        this.errMsg = 'Maximum number must be a number !!'; 
     }
     else {
       this.errorMessageDuringAdd = false ;
@@ -68,6 +73,7 @@ export class DoctorScheduleComponent implements OnInit , OnDestroy {
         this.workIn.push(this.work);
         this.showResult = true;
         this.showAddToSchedule= false ;
+        this.info= {day:"" , start:"" ,sShift :"" , end:"" , eShift :"" ,maxNo:""};
 
       },error:err=>{
         console.log(err);
@@ -95,7 +101,7 @@ export class DoctorScheduleComponent implements OnInit , OnDestroy {
           this.doctor = data[0].didNavigation ;
           this.showResult= true ;
     },error:err=>{
-
+      console.log(err);
     }})
    }
 
@@ -105,14 +111,16 @@ export class DoctorScheduleComponent implements OnInit , OnDestroy {
    }
 
    deleteRow(index:number , sTime :string, did:number){
-      console.log(index,sTime,did);
-      this.subscribe = this.adminService.deleteRowFromSchedule(did,sTime).subscribe({next:data=>{
-            console.log(data);
-            this.workIn.splice(index,1);
-            alert('Deleted successfuly ')
-      },error:err=>{
-        console.log(err);
-      }})
-   }
+      //console.log(index,sTime,did);
+      var res = confirm("Are you sure that you want to delete this record ?");
+      if(res){
+        this.subscribe = this.adminService.deleteRowFromSchedule(did,sTime).subscribe({next:data=>{
+          console.log(data);
+          this.workIn.splice(index,1);
+          alert('Deleted successfuly');
+        },error:err=>{
+          console.log(err);
+        }})
+      }
+    } 
 }
-
